Login acces modified
updated the programe regulate the user's access to some page depending on whether the user is logged in or not. if the user is logged in he can no longer access the login, register or forget password pages. if not logged in he will access no other pages but the ones i mentioned earlier.